requirements.txt
streamlit
plotly
requests
pandas
numpy
xgboost
scikit-learn
python-dotenv
transformers
torch